+++ GaggiaPIDController/Sensors.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "Configuration.h"
+
 #include "max6675.h"
 #include <TSIC.h>