+++ GaggiaPIDController/SerialInterface.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "Common.h"
-#include <Arduino.h>
 
 struct SerialInput
 {
@@ -13,25 +12,155 @@ struct SerialInput
     double kd = 0.0;
 };
 
-class SerialInterface
+template <class Adapter> class SerialInterface
 {
   public:
-    SerialInterface(const uint16_t &baudrate);
+    SerialInterface(const unsigned short &baudrate)
+    {
+        Adapter::SerialBegin(baudrate);
+    }
 
-    void read_input();
+    void read_input()
+    {
+        using string = typename Adapter::String;
 
-    void print_status(const Gaggia::ControlStatus &status);
-    void print_help();
+        // Read input data and enable or disable the debug mode
+        // If debug mode is enabled, accept temperature input
+        // to overwrite sensor readings
+        if (Adapter::SerialAvailable())
+        {
+            typename Adapter::String data = Adapter::SerialReadStringUntil('\n');
+            if (data.startsWith("help"))
+            {
+                print_help();
+            }
+            else if (data.startsWith("debug on"))
+            {
+                inputs.debug_mode = true;
+                Adapter::SerialPrintln("Setting debug mode ON");
+            }
+            else if (data.startsWith("debug off"))
+            {
+                inputs.debug_mode = false;
+                Adapter::SerialPrintln("Setting debug mode OFF");
+            }
+            else if (data.startsWith("output on"))
+            {
+                inputs.enable_output = true;
+                Adapter::SerialPrintln("Setting output ON");
+            }
+            else if (data.startsWith("output off"))
+            {
+                inputs.enable_output = false;
+                Adapter::SerialPrintln("Setting output OFF");
+            }
+            else if (data.startsWith("temp ") && is_debug_active())
+            {
+                data.replace("temp ", "");
+                inputs.mock_temperature = data.toDouble();
+                Adapter::SerialPrintln("Setting mock temperature to; " + data);
+            }
+            else if (data.startsWith("kp "))
+            {
+                data.replace("kp ", "");
+                inputs.kp = data.toDouble();
+                Adapter::SerialPrintln("Setting PID kp to: " + data);
+            }
+            else if (data.startsWith("ki "))
+            {
+                data.replace("ki ", "");
+                inputs.ki = data.toDouble();
+                Adapter::SerialPrintln("Setting PID ki to: " + data);
+            }
+            else if (data.startsWith("kd "))
+            {
+                data.replace("kd ", "");
+                inputs.kd = data.toDouble();
+                Adapter::SerialPrintln("Setting PID kd to: " + data);
+            }
+        }
+    }
 
-    bool is_debug_active();
-    double get_mock_temperature();
-    bool is_output_enabled();
-    bool get_new_kp(double *kp);
-    bool get_new_ki(double *ki);
-    bool get_new_kd(double *kd);
+    void print_status(const Gaggia::ControlStatus<Adapter> &status)
+    {
+        using string = typename Adapter::String;
+        auto now = Adapter::millis();
+        if (is_output_enabled() && now - time_last_print > PRINT_TIMEOUT)
+        {
+            time_last_print = now;
+            string output = string(static_cast<int>(status.machine_mode)) + "," +
+                            string(status.current_temperature) + "," +
+                            string(status.target_temperature) + "," +
+                            string(status.water_heater_on) + "," + status.status_message;
+            Adapter::SerialPrintln(output);
+        }
+    }
+
+    void print_help()
+    {
+        Adapter::SerialPrintln("**** GaggiaPIDController serial interface ****");
+        Adapter::SerialPrintln("Commands:");
+        Adapter::SerialPrintln("- help: show this message");
+        Adapter::SerialPrintln("- debug [on/off]: enable/disable debug mode");
+        Adapter::SerialPrintln("- output [on/off]: enable/disable serial output");
+        Adapter::SerialPrintln(
+            "- temp [value]: Set the mock temperature. Used only in debug mode");
+        Adapter::SerialPrintln("- kp [value]: Set the P gain of the PID controller");
+        Adapter::SerialPrintln("- ki [value]: Set the I gain of the PID controller");
+        Adapter::SerialPrintln("- kd [value]: Set the D gain of the PID controller");
+        Adapter::SerialPrintln("**** END ****");
+    }
+
+    bool is_debug_active()
+    {
+        return inputs.debug_mode;
+    }
+
+    double get_mock_temperature()
+    {
+        return inputs.mock_temperature;
+    }
+
+    bool is_output_enabled()
+    {
+        return inputs.enable_output;
+    }
+
+    bool get_new_kp(double *kp)
+    {
+        if (inputs.kp > 0.0)
+        {
+            *kp = inputs.kp;
+            inputs.kp = 0.0; // reset input
+            return true;
+        }
+        return false;
+    }
+
+    bool get_new_ki(double *ki)
+    {
+        if (inputs.ki > 0.0)
+        {
+            *ki = inputs.ki;
+            inputs.ki = 0.0; // reset input
+            return true;
+        }
+        return false;
+    }
+
+    bool get_new_kd(double *kd)
+    {
+        if (inputs.kd > 0.0)
+        {
+            *kd = inputs.kd;
+            inputs.kd = 0.0; // reset input
+            return true;
+        }
+        return false;
+    }
 
   private:
-    static constexpr uint16_t PRINT_TIMEOUT = 200;
+    static constexpr unsigned short PRINT_TIMEOUT = 200;
     unsigned long time_last_print;
     SerialInput inputs;
 };