+++ GaggiaPIDController/TemperatureSensor.h
@@ -1,14 +1,11 @@
 #pragma once
 
-#include "Common.h"
-#include <Arduino.h>
-
-class BaseTemperatureSensor
+template <class Adapter> class BaseTemperatureSensor
 {
   public:
     /** Return the name identifier of the sensor.
      */
-    virtual String get_name() = 0;
+    virtual typename Adapter::String get_name() = 0;
 
     /* Read the sensor and store the current temperature in
      * celsius degrees into 'value'.
@@ -16,24 +13,28 @@ class BaseTemperatureSensor
     virtual bool get_temperature_celsius(float *value) = 0;
 };
 
-template <class SensorType> class TemperatureSensor : public BaseTemperatureSensor
+template <class Adapter, class SensorType>
+class TemperatureSensor : public BaseTemperatureSensor<Adapter>
 {
   public:
-    TemperatureSensor(const String &name, const uint8_t &pin, const uint32_t read_period,
-                      const uint32_t &moving_avg_size = 1, const float &temp_offset = 0)
-        : name(name), m_avg(moving_avg_size), time_last_read(millis()), healthy(true),
-          read_period(read_period), temp_offset(temp_offset), sensor(pin)
+    TemperatureSensor(const typename Adapter::String &name,
+                      const typename Adapter::uint8_t &pin,
+                      const typename Adapter::uint32_t read_period,
+                      const typename Adapter::uint32_t &moving_avg_size = 1,
+                      const float &temp_offset = 0)
+        : name(name), m_avg(moving_avg_size), time_last_read(Adapter::millis()),
+          healthy(true), read_period(read_period), temp_offset(temp_offset), sensor(pin)
     {
     }
 
-    String get_name()
+    typename Adapter::String get_name()
     {
         return name;
     }
 
     bool get_temperature_celsius(float *value)
     {
-        unsigned long now = millis();
+        unsigned long now = Adapter::millis();
         if (now - time_last_read > read_period)
         {
             time_last_read = now;
@@ -54,9 +55,9 @@ template <class SensorType> class TemperatureSensor : public BaseTemperatureSens
 
   private:
     SensorType sensor;
-    String name;
+    typename Adapter::String name;
     unsigned long time_last_read;
-    uint32_t read_period;
+    typename Adapter::uint32_t read_period;
     MovingAverage<float> m_avg;
     bool healthy;
     float temp_offset;